Update install.sh
Updated the install logic inline with the aarch64 version as it was much more sound and didnt fail when it couldnt find an existing xbin directory on occasion on some ROM's. Did i ever mention bash is not my natural habitat? Ive been suckling at the teat of Microsoft for too long.....need to get back to my roots more :)